Support Content-Security-Policies via nonce values
Content-Security-Policy is a common way of restricting which assets
(javascript, styles, etc) are loaded by the browser as a way of
preventing a number of vulnerabilities.

Flask-Admin loads a fair amount of JavaScript and styling to support its
functionality, and previously it has been difficult to get this to pass
CSP checks.

We can add support for CSP by taking a CSP nonce generator and using
that to inject a nonce value wherever we load these resources (either
asking the client to download files or with inline scripts/styles).